Move utils.py into biodatasets/
